fix: undefined문제 옵셔널 체이닝 사용, 불필요한 구문 삭제, utils 폴더 추가, useEffect 안에 함수 선언
