+++ src/components/App.css
@@ -39,6 +39,7 @@ body {
 
 a {
   text-decoration: none;
+  cursor: pointer;
 }
 
 p {
