+++ src/utils/formatDate.js
@@ -1,0 +1,6 @@
+export default function formatDate(date) {
+  const formattedDate = new Date(date);
+  return `${formattedDate.getFullYear()}. ${
+    formattedDate.getMonth() + 1
+  }. ${formattedDate.getDate()}`;
+}
